refactor(dwaar-cli): extract shared fetch_latest_version helper (#176)
self_update and auto_update both queried the GitHub releases API
to find the latest version tag. self_update had a richer
JSON-then-redirect fallback path; auto_update only did the redirect.
Extract the robust version into version_check::fetch_latest_version
and call it from both. Tag-parsing tests move with the helper.

Closes #176
